move CacheKey from BaseGVFSOperations to FileSystemUtils
